requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
Pillow>=10.0.0
hydra-core>=1.3.2
omegaconf>=2.3.0
wandb>=0.16.0
datasets>=2.16.0
tqdm>=4.66.0
matplotlib>=3.8.0
torchmetrics[image]>=1.3.0
pytest>=8.0.0
ruff>=0.3.0
pre-commit>=3.6.0
